chore: convert pipedproviders to TypeScript
