+++ package.json
@@ -92,8 +92,6 @@
     "cookie-parser": "^1.4.3",
     "cors": "^2.5.2",
     "debug": "^4.3.3",
-    "deep-get-set": "^1.1.0",
-    "dev-null-stream": "0.0.1",
     "devcert": "^1.2.2",
     "dnssd2": "1.0.0",
     "errorhandler": "^1.3.0",
